Thu, Aug 17, 2023, 12:05 PM +06:00

+++ beecrowd/1551.c
@@ -25,7 +25,7 @@ int main(void)
 
             if(str[j] >= 'a' && str[j] <= 'z')
             {
-                freq[str[j]] = 1;
+                freq[(int)str[j]] = 1;
             }
             j++;
         }
